exposed mapper in a separate module, fix #632
