+++ examples/example/src/main/java/org/camunda/community/rest/example/standalone/client/ProcessClient.java
@@ -6,7 +6,7 @@ import org.camunda.bpm.engine.repository.ProcessDefinition;
 import org.camunda.bpm.engine.runtime.MessageCorrelationResultWithVariables;
 import org.camunda.bpm.engine.runtime.ProcessInstance;
 import org.camunda.bpm.engine.variable.VariableMap;
-import org.camunda.community.rest.variables.PrettyPrinter;
+import org.camunda.community.rest.util.PrettyPrinter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
