+++ extension/variables/src/license/licenses.properties
@@ -1,0 +1,1 @@
+camunda_apache_v2=Camunda Apache 2 License
