+++ extension/variables/src/main/kotlin/org/camunda/community/rest/variables/ValueTypeResolverImpl.kt
@@ -3,7 +3,6 @@ package org.camunda.community.rest.variables
 import org.camunda.bpm.engine.variable.type.ValueType
 import org.camunda.bpm.engine.variable.type.ValueTypeResolver
 import org.springframework.stereotype.Component
-import java.util.*
 
 /**
  * Implementation of the [ValueTypeResolver] that is taken from the Camunda engine.
@@ -48,24 +47,4 @@ class ValueTypeResolverImpl : ValueTypeResolver {
 
     return types
   }
-
-
-}
-
-/**
- * Tries to guess the type from the passed value.
- */
-@Suppress("PLATFORM_CLASS_MAPPED_TO_KOTLIN")
-fun Any?.resolveValueType(): ValueType = when (this) {
-  null -> ValueType.NULL
-  is Boolean -> ValueType.BOOLEAN
-  is Date -> ValueType.DATE
-  is Double -> ValueType.DOUBLE
-  is Integer -> ValueType.INTEGER
-  is Long -> ValueType.LONG
-  is Short -> ValueType.SHORT
-  is String -> ValueType.STRING
-  is ByteArray -> ValueType.BYTES
-  is Number -> ValueType.NUMBER
-  else -> ValueType.OBJECT
 }
