+++ pom.xml
@@ -59,6 +59,7 @@
 
   <modules>
     <module>extension/client</module>
+    <module>extension/variables</module>
     <module>extension/core</module>
     <module>extension/starter</module>
     <module>extension/starter-provided</module>
@@ -93,6 +94,13 @@
         <version>${camunda.version}</version>
         <type>pom</type>
         <scope>import</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.camunda.commons</groupId>
+        <artifactId>camunda-commons-bom</artifactId>
+        <version>${camunda.version}</version>
+        <scope>import</scope>
+        <type>pom</type>
       </dependency>
       <dependency>
         <groupId>io.holunda.camunda-api</groupId>
